client/requirements.txt
google-genai
pydantic
psycopg2-binary
requests
mcp
fastapi
uvicorn
streamlit
groq
openai